requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8
pre-commit=2.20.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
